Atualizando tamanho para indice de leitura

+++ include/file_interface.hpp
@@ -14,7 +14,7 @@ class FileInterface {
     void write(char* bytes, unsigned size);
     void write(char* bytes, unsigned size, unsigned long long int position);
     char* read(unsigned size);
-    char* read(unsigned size, unsigned position);
+    char* read(unsigned size, unsigned long long int position);
     
     void close();
     bool isOpen();
